+++ Movies.Aggregates/Film/MovieRepositoryBase.cs
@@ -1,0 +1,6 @@
+﻿namespace Movies.Aggregates.Film
+{
+	public class MovieRepositoryBase
+	{
+	}
+}