+++ Movies.Contracts/IMovieGrain.cs
@@ -1,11 +1,16 @@
-﻿using Orleans;
+﻿using Movies.Contracts.Entity;
+using Orleans;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace Movies.Contracts
 {
 	public interface IMovieGrain : IGrainWithStringKey
 	{
-		Task<MovieModel> Get();
-		Task Set(string name);
+		Task<MovieModel> Get(string id);
+		Task<IEnumerable<MovieModel>> GetList(string genre);
+		Task<IEnumerable<MovieModel>> GetRatedMovies();
+		Task Update(string id, MovieModel entity);
+		Task Set(MovieModel entity);
 	}
 }