+++ Movies.Contracts/IMovieGrainClient.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using Movies.Contracts.Entity;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using Zeppeling.Framework.Abstactions.Response;
 
@@ -6,10 +7,10 @@ namespace Movies.Contracts
 {
 	public interface IMovieGrainClient
 	{
-		Task<MovieModel> Get(string id);
+		Task<Entity.MovieModel> Get(string id);
 		Task<IEnumerable<MovieModel>> GetList(string genre);
-		Task<IEnumerable<MovieModel>> GetRatedFilms();
-		Task Set(string key, string name);
-		Task Update(string key, string name);
+		Task<IEnumerable<MovieModel>> GetRatedMovies();
+		Task Set(MovieModel entity);
+		Task Update(string id, MovieModel entity);
 	}
 }