+++ Movies.GrainClients/GrainClientsServiceCollectionExtensions.cs
@@ -1,10 +1,14 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
+using Movies.Aggregates.Film;
 using Movies.Contracts;
 
 namespace Movies.GrainClients
 {
 	public static class GrainClientsServiceCollectionExtensions
 	{
-		public static void AddAppClients(this IServiceCollection services) => services.AddSingleton<IMovieGrainClient, MovieGrainClient>();
+		public static void AddAppClients(this IServiceCollection services) {
+			services.AddScoped<IMovieRepository, MovieRepository>();
+			services.AddScoped<IMovieGrainClient, MovieGrainClient>();
+		} 
 	}
 }