+++ Movies.Grains/MovieGrain.cs
@@ -1,6 +1,9 @@
-﻿using Movies.Contracts;
+﻿using Movies.Aggregates.Film;
+using Movies.Contracts;
+using Movies.Contracts.Entity;
 using Orleans;
 using Orleans.Providers;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace Movies.Grains
@@ -8,12 +11,41 @@ namespace Movies.Grains
 	[StorageProvider(ProviderName = "Default")]
 	public class MovieGrain : Grain<MovieModel>, IMovieGrain
 	{
-		public Task<MovieModel> Get()
-			=> Task.FromResult(State);
+		private IMovieRepository _repository;
 
-		public Task Set(string name)
+		public async Task<MovieModel> Get(string id)
 		{
-			//State = new MovieModel(this.GetPrimaryKeyString(),name);
+			_repository = new MovieRepository();
+			MovieModel response = await _repository.Get(id);
+			return response;
+		}
+
+		public async Task<IEnumerable<MovieModel>> GetList(string genre)
+		{
+			_repository = new MovieRepository();
+			IEnumerable<MovieModel> response = await _repository.GetList(genre);			
+
+			return response;
+		}
+
+		public async Task<IEnumerable<MovieModel>> GetRatedMovies() 
+		{
+			_repository = new MovieRepository();
+			IEnumerable<MovieModel> response = await _repository.GetList(string.Empty);
+			return response;
+		}
+
+		public Task Set(MovieModel entity)
+		{
+			_repository = new MovieRepository();
+			Task.FromResult(_repository.AddAsync(entity));
+			return Task.CompletedTask;
+		}
+
+		public Task Update(string id, MovieModel entity)
+		{
+			_repository = new MovieRepository();
+			Task.FromResult(_repository.UpdateAsync(id, entity));
 			return Task.CompletedTask;
 		}
 	}