+++ Movies.Repository.Test/UnitTest1.cs
@@ -1,0 +1,13 @@
+using Xunit;
+
+namespace Movies.Repository.Test
+{
+    public class UnitTest1
+    {
+        [Fact]
+        public void Test1()
+        {
+
+        }
+    }
+}