+++ Movies.Server/Gql/Types/MovieGraphType.cs
@@ -1,5 +1,7 @@
 ﻿using GraphQL.Types;
+using Movies.Aggregates.Film;
 using Movies.Contracts;
+using Movies.Contracts.Entity;
 using Zeppeling.Framework.Abstactions.Response;
 
 namespace Movies.Server.Gql.Types
@@ -11,14 +13,14 @@ namespace Movies.Server.Gql.Types
 			Name = "Movie";
 			Description = "Movie Data Type";
 
-			Field(x => x.Id, nullable: true).Description("Unique key.");
-			Field(x => x.Name, nullable: true).Description("Name.");
-			Field(x => x.Key, nullable: true).Description("Key.");
-			Field(x => x.Description, nullable: true).Description("Description.");
-			Field(x => x.Genres, nullable: true).Description("Genres.");
-			Field(x => x.Rate, nullable: true).Description("Rate.");
-			Field(x => x.Length, nullable: true).Description("Length.");
-			Field(x => x.Img, nullable: true).Description("Img.");
+			Field(x => x._id, nullable: true).Description("Unique key.");
+			Field(x => x.name, nullable: true).Description("Name.");
+			Field(x => x.key, nullable: true).Description("Key.");
+			Field(x => x.description, nullable: true).Description("Description.");
+			// Field(x => x.genres, nullable: true).Description("Genres.");
+			Field(x => x.rate, nullable: true).Description("Rate.");
+			Field(x => x.length, nullable: true).Description("Length.");
+			Field(x => x.img, nullable: true).Description("Img.");
 		}
 	}
 }